{{cookiecutter.project_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name}}"
version = "0.1.0"
description = ""
authors = ["{{cookiecutter.author_name}} <{{cookiecutter.author_email}}>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.0.7"
drf-oidc-auth = "^0.10.0"
djangorestframework = "^3.11.0"
markdown = "^3.2.2"
django-filter = "^2.3.0"
drf-nested-routers = "^0.91"
psycopg2 = "^2.8.5"
python-dotenv = "^0.13.0"
django-cors-headers = "^3.4.0"
django-grappelli = "^2.15.1"
gunicorn = "^20.0.4"
django-split-settings = "^1.0.1"
django-postgres-extra = "^1.22"
django-health-check = "^3.16.4"

[tool.poetry.dev-dependencies]
pre-commit = "^2.5.1"
django-debug-toolbar = "^3.2.1"
pytest = "^6.2.4"
pytest-django = "^4.4.0"
flake8 = "^3.9.2"
black = "^19.10b0"
reorder-python-imports = "^2.3.1"
pyflakes = "^2.3.1"
pycodestyle = "^2.7.0"

[tool.black]
line-length = 95

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
